+++ src/handlesProcesses/handleDel.js
@@ -1,0 +1,9 @@
+import deleteSymbols from './utils/deleteSymbols';
+
+const handleDel = (textField) => {
+  const { selectionStart } = textField;
+
+  deleteSymbols(textField, selectionStart);
+};
+
+export default handleDel;
