+++ src/handlesProcesses/handleLeftArrow.js
@@ -1,0 +1,8 @@
+const handleLeftArrow = (textField) => {
+  const { selectionStart } = textField;
+  const cursorPosition = selectionStart === 0 ? 0 : selectionStart - 1;
+
+  textField.setSelectionRange(cursorPosition, cursorPosition);
+};
+
+export default handleLeftArrow;
