+++ src/handlesProcesses/handleRightArrow.js
@@ -1,0 +1,9 @@
+const handleRightArrow = (textField) => {
+  const { selectionStart, value } = textField;
+  const { length } = value;
+  const cursorPosition = selectionStart === length ? length : selectionStart + 1;
+
+  textField.setSelectionRange(cursorPosition, cursorPosition);
+};
+
+export default handleRightArrow;
