--- src/renders/setLang.js
@@ -1,8 +1,0 @@
-const setLang = (lang) => {
-  const locale = { lang };
-  localStorage.setItem('locale', JSON.stringify(locale));
-  // const currentLocale = JSON.parse(localStorage.getItem('locale'));
-  // state.lang = currentLocale.lang;
-};
-
-export default setLang;
